feat: create CollectionEditor component
